requirements.txt
# aws-cdk-lib==2.53.0
# constructs>=3.3.69,<4.0.0
aws-cdk.aws_s3
aws-cdk.core